+++ sdk/crypto/jose/json_web_key.hpp
@@ -58,7 +58,7 @@ class json_web_key : public crypto_keychain {
      * @param crypto_key* cryptokey [in]
      * @param keyflag_t mode [in] see keyflag_t
      * @param stream_t* stream [in]
-     * @param int flag [in] key_public, key_private
+     * @param int flags [in] public_key | private_key
      * @return error code (see error.hpp)
      * @example
      *          json_web_key jwk;
@@ -68,21 +68,21 @@ class json_web_key : public crypto_keychain {
      *          bin.resize (size);
      *          jwk.write (&privkey, &bin[0], &size);
      */
-    virtual return_t write(crypto_key* cryptokey, keyflag_t mode, stream_t* stream, int flag = 0);
+    virtual return_t write(crypto_key* cryptokey, keyflag_t mode, stream_t* stream, int flags = public_key);
     /**
      * @brief   write
      * @param   crypto_key* cryptokey [in]
      * @param   std::string& buf [out]
-     * @param   int flag [inopt] 0 public only, 1 also private
+     * @param   int flags [inopt] public_key | private_key
      */
-    return_t write(crypto_key* cryptokey, std::string& buf, int flag = 0);
+    return_t write(crypto_key* cryptokey, std::string& buf, int flags = public_key);
     /**
      * @brief   write
      * @param   crypto_key* cryptokey [in]
      * @param   stream_t* buf [out]
-     * @param   int flag [inopt] 0 public only, 1 also private
+     * @param   int flags [inopt] public_key | private_key
      */
-    return_t write(crypto_key* cryptokey, stream_t* buf, int flag = 0);
+    return_t write(crypto_key* cryptokey, stream_t* buf, int flags = public_key);
 
    protected:
     return_t read_json_keynode(crypto_key* cryptokey, json_t* json);