+++ CHANGELOG.md
@@ -4,6 +4,9 @@
   * [study] QUIC (RFC 9001, 9369)
   * [study] TLS (RFC 4346, 5246, 8446, 8448)
 
+* Revision 717
+  * [fixed] json_web_key, cbor_web_key write member
+
 * Revision 716
   * [tested] RFC 9369 QUIC Version 2